refactor: extract domain utilities and constants into distinct modules
