+++ packages/core/src/common/constants.ts
