+++ packages/core/src/utils/index.ts
@@ -1,1 +1,3 @@
 export * from "./circular-buffer";
+export * from "./k-weighting";
+export * from "./loudness-math";
